+++ src/main/java/com/openrca/ecommerce/repository/ShipRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShipRepository extends JpaRepository<Ship, Integer> {
+public interface ShipRepository extends JpaRepository<Ship, Long> {
 }
